chore: minor refactor
