+++ client-app/shared/account/components/account-navigation.vue
@@ -24,7 +24,7 @@
 import { computed } from "vue";
 import { useNavigations } from "@/core/composables";
 import { useUser } from "@/shared/account/composables/useUser";
-import { useExtensionRegistry } from "@/shared/common/composables/useExtensionRegistry";
+import { useExtensionRegistry } from "@/shared/common/composables/extensionRegistry/useExtensionRegistry";
 import LinkDefault from "./account-navigation-link-components/link-default.vue";
 import type { ExtendedMenuLinkType } from "@/core/types";
