+++ client-app/shared/common/composables/extensionRegistry/useExtensionRegistry.ts
